+++ bopep/bayesian_optimization/optimization.py
@@ -15,8 +15,9 @@ from bopep.surrogate_model import (
 )
 from bopep.logging.logger import Logger
 from bopep.bayesian_optimization.acquisition_functions import AcquisitionFunction
-from bopep.bayesian_optimization.utils import (_check_binding_site_residue_indices, _validate_checkpoint, _validate_dependencies, _validate_args, _validate_surrogate_model_kwargs)
-from bopep.bayesian_optimization.checkpointing import _next_checkpoint_dir, _save_checkpoint, _copy_logs_to_checkpoint, _setup_checkpoint_dir, _rebuild_logs_from_csvs
+from bopep.bayesian_optimization.utils import (_validate_dependencies, _validate_args, _validate_surrogate_model_kwargs)
+from bopep.bayesian_optimization.pdb_utils import _check_binding_site_residue_indices
+from bopep.bayesian_optimization.checkpointing import _next_checkpoint_dir, _save_checkpoint, _copy_logs_to_checkpoint, _setup_checkpoint_dir, _rebuild_logs_from_csvs, _validate_checkpoint
 from bopep.bayesian_optimization.train_validate_utils import _compute_model_metrics, _compute_split_indices, _train_and_validate, _split_train_validation, _train_on_all
 from bopep.bayesian_optimization.selection import PeptideSelector
 from bopep.scoring.scores_to_objective import ScoresToObjective
